Discontinue use of results/buildlogs directory entirely
Upon further reflection, once we dispose of the idea of just one run of
cpanm being performed so as to get a build.log which is then parsed to
get many .json files which are the basis of analysis, there is no longer
need for retaining the build.log files anywhere (except where cpanm
itself creates them, i.e., in
.../testing/bin/.cpan/work/<timestamped.directory>/build.log).  If I
were to gzip each build.log file before depositing it in
.../results/buildlogs/, what would I call it?  The timestamp doesn't
give a clue as to what modules -- and their prerequisites -- were parsed
in any given build.log.  All the useful data has been extracted into a
.log.json file and those files have been deposited in
.../results/analysis/ named in a useful way.  So I can dispense with
results/storeage/ entirely.  The files I was placing in
.../results/buildlogs serve no useful purpose, whether gzipped or not.

This has the further implication that I don't need to develop a list of
buildlogs as process_modules() runs, nor do I need to return
$buildlogs_ref from that method, nor do I need to offer a
retain_buildlogs switch to the final method.

Edit Commit.pm to remove code and documentation.  Remove tests from
t/003.  (Over 200 lines of code or docs or tests removed!)